re_contact_book_system/backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic[email]==2.10.4
email-validator==2.2.0

# Utils
python-dotenv==1.0.1
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.4
httpx==0.28.1
